staff_hubのモデルを定義しなおし
