+++ backend/apps/staff_hub/models/user.py
@@ -2,21 +2,7 @@ from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin, BaseU
 from django.db import models
 from apps.utility.const import MESSAGES
 from apps.utility.models import BaseModel
-
-class Organization(BaseModel):
-    """
-    組織モデル
-    """
-    organization_name = models.CharField("組織名", max_length=255, unique=True)
-    description = models.TextField("組織説明", blank=True, null=True)
-
-    class Meta:
-        verbose_name = '組織'
-        verbose_name_plural = '組織'
-        db_table = 'organizations'
-
-    def __str__(self):
-        return self.organization_name
+from apps.staff_hub.models.organization import Organization
 
 
 class UserManager(BaseUserManager):
@@ -69,27 +55,3 @@ class User(AbstractBaseUser, PermissionsMixin, BaseModel):
     def __str__(self):
         return self.email
 
-
-class Permission(BaseModel):
-    """
-    権限モデル
-    """
-    user = models.OneToOneField(User, on_delete=models.CASCADE, verbose_name="ユーザー")
-    material_access = models.BooleanField("資料アクセス：全体管理", default=False)
-    can_manage_own_attendance = models.BooleanField("勤怠アクセス：自分のみ管理", default=False)
-    can_manage_all_attendance = models.BooleanField("勤怠アクセス：全体管理", default=False)
-    can_view_production_plan = models.BooleanField("生産計画アクセス：閲覧", default=False)
-    can_edit_production_plan = models.BooleanField("生産計画アクセス：編集・削除", default=False)
-    can_view_order = models.BooleanField("受注アクセス：閲覧", default=False)
-    can_edit_order = models.BooleanField("受注アクセス：編集・削除", default=False)
-    can_view_defect = models.BooleanField("不具合アクセス：閲覧", default=False)
-    can_edit_defect = models.BooleanField("不具合アクセス：編集・削除", default=False)
-    mail_access = models.BooleanField("メールアクセス：全体管理", default=False)
-
-    class Meta:
-        verbose_name = 'アクセス権限'
-        verbose_name_plural = 'アクセス権限'
-        db_table = 'permissions'
-
-    def __str__(self):
-        return self.user.name
