+++ backend/apps/staff_hub/views.py
@@ -1,7 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
-from .models import Permission
+from .models.permission import Permission
 from .serializers import CustomUserSerializer
 
 class CustomUserMeView(APIView):
