+++ backend/tests/factory/permission_factory.py
@@ -1,10 +1,6 @@
 import factory
-from apps.staff_hub.models import Permission
+from apps.staff_hub.models.permission import Permission
 from .user_factory import UserFactory
-
-import factory
-from apps.staff_hub.models import Permission
-from tests.factory.user_factory import UserFactory
 
 
 class PermissionFactory(factory.django.DjangoModelFactory):
